camerad: camera_util is spectra (#33597)
* camerad: camera_util is spectra

* lil more

* mv that

* fix build

* runs on device

---------
